uwsgi: fix freeze, http/1.1 and keepalive
Sometimes connection to gnocchi-api freeze because the body is not fully
read. This may due to a Gnocchi returing an error before the wsgi.input
is fully read, or a wrong/missing content-length.

This change ensures we always read the entire wsgi.input with a Pecan after
hook.

And it switches back uwsgi router to "http" to ensure uwsgi acts as a
HTTP Server (ensure correctness of the HTTP protocol, before passing the
data to webob). Webob will always have the correct content-length.

This also allows to re-enable keepalive of Gnocchi API by default.

But this have a bad side-effect, the influxdb ingestion will not work
out-of-the-box because it uses chunked enconding, and this is not
supported with the uwsg "http" router [1].

To work around this, a new option "[api]/uwsgi_mode" allow to switch
back to http-socket, with some doc to explain this mode requires a real
HTTP Server on front of Gnocchi API for HTTP protocol correctness.

[1] https://github.com/unbit/uwsgi/issues/1428

Closes-bug: #156